This commit introduces a major change, separating the ethoscope activities in two services:
ethoscope_listener 	controls the core ethoscope activities (tracking, video recording, streaming) agnostic of the network
ethoscope_device	controls the web server running on port 9000 which then is used to interface the ethoscope controls with the node

These two can now run as different services. This offers more flexibility in case of network problems. For instance, if something happens to the network it is possible to restart ethoscope_device to re-register the ethoscope without affecting any underlying activity.
The code should generally be slimmer and more robust.

However, for this change to take place ethoscope_listener.service needs to be copied to /usr/lib/systemd/system and enabled via systemctl
Until this is done, ethoscope_device will start the service automatically (thus defeating the purpose of this change).